Add get_executable_name() function
This function wraps rcutils_get_executable_name and takes care of the
allocator and throws an exception when an error occurs.

The rcutils_get_executable_name implementation doesn't set any error
information when it fails, and from reading through the implementation
and the documentation for the system functions it calls, an error should
be extremely unlikely. A generic failure message is the best I can do.

Signed-off-by: Scott K Logan <logans@cottsay.net>