Add an ASSERT to the pointer traits tests.
What's happening here is that we are using the initialization
of the pointer to initialize the "auto" type of the variable,
and then using compile-time checks to check on it.  Because
of this, clang static analysis thinks that the variable is
unused (which it is, at runtime).  Since that is not the
purpose of the test, just put a bogus ASSERT_TRUE in that uses
the variable and quiets clang static analysis.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>