Fix working with filesystem parent paths.
I found two separate bugs while working with parent paths:

1.  Asking for the first parent path of an absolute path
would add a duplicate separator at the beginning.  I fixed
this by checking to see if the separator already exists,
and skiping if there is already one.
2.  Asking for the grandparent path of an absolute path
would return a relative path, which is very incorrect.  The
problem here turned out to be that we were not propagating
empty vector pieces from a child to a parent, and so the
grandparent would miss out that it was supposed to be absolute.
Fix this by always adding all pieces to the parents, even
when they are empty.

I also added tests for the above cases.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>