Fix working with filesystem parent paths. (#112)
* Fix working with filesystem parent paths.

I fixed a few problems found when working with parent paths:

1.  Asking for the first parent path of an absolute path
would add a duplicate separator at the beginning.  I fixed
this by checking to see if the separator already exists,
and skipping if there is already one.
2.  Asking for the grandparent path of an absolute path
would return a relative path, which is very incorrect.  The
problem here turned out to be that we were not propagating
empty vector pieces from a child to a parent, and so the
grandparent would miss out that it was supposed to be absolute.
Fix this by always adding all pieces to the parents, even
when they are empty.
3.  Added tests for the above situations.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>